integrate update menu input view
