+++ EngineerMemo/Presentation/Component/CustomStyle/CustomStyle.swift
@@ -1,7 +1,7 @@
 import UIKit
 import UIKitHelper
 
-// MARK: - 入力
+// MARK: - 入力(View)
 
 extension ViewStyle where T: UIView {
     static var inputView: ViewStyle<T> {
@@ -21,6 +21,67 @@ extension ViewStyle where T: UIView {
             $0.layer.borderColor = UIColor.primary.cgColor
             $0.layer.borderWidth = 1.0
             $0.clipsToBounds = true
+        }
+    }
+}
+
+// MARK: - 入力(ボタン)
+
+extension ViewStyle where T: UIButton {
+    static func menuButton(
+        title: String?,
+        actions: [UIMenuElement]
+    ) -> ViewStyle<T> {
+        .init {
+            var config = UIButton.Configuration.filled()
+            config.title = title
+            config.baseForegroundColor = .primary
+            config.contentInsets = .init(top: 0, leading: 8, bottom: 0, trailing: 0)
+            config.titleTextAttributesTransformer = .init { incoming in
+                var outgoing = incoming
+                outgoing.font = .systemFont(ofSize: 16)
+                return outgoing
+            }
+            config.background.backgroundColor = .background
+            $0.configuration = config
+            $0.contentHorizontalAlignment = .leading
+            $0.showsMenuAsPrimaryAction = true
+            $0.menu = .init(
+                title: .empty,
+                options: .displayInline,
+                children: actions
+            )
+        }
+    }
+
+    static func menuButton(
+        title: String?,
+        icon: UIImage?,
+        actions: [UIMenuElement]
+    ) -> ViewStyle<T> {
+        .init {
+            var config = UIButton.Configuration.filled()
+            config.title = title
+            config.image = icon?
+                .resized(size: .init(width: 24, height: 24))
+                .withRenderingMode(.alwaysOriginal)
+            config.baseForegroundColor = .primary
+            config.contentInsets = .init(top: 0, leading: 8, bottom: 0, trailing: 0)
+            config.imagePadding = 8
+            config.titleTextAttributesTransformer = .init { incoming in
+                var outgoing = incoming
+                outgoing.font = .boldSystemFont(ofSize: 16)
+                return outgoing
+            }
+            config.background.backgroundColor = .background
+            $0.configuration = config
+            $0.contentHorizontalAlignment = .leading
+            $0.showsMenuAsPrimaryAction = true
+            $0.menu = .init(
+                title: .empty,
+                options: .displayInline,
+                children: actions
+            )
         }
     }
 }
