+++ EngineerMemo/Presentation/Screen/Memo/Update/MemoUpdateContentView.swift
@@ -5,12 +5,7 @@ import UIKitHelper
 // MARK: - properties & init
 
 final class MemoUpdateContentView: UIView {
-    @Published private(set) var selectedCategoryType: MemoInputCategoryType? {
-        didSet {
-            setupCategory()
-        }
-    }
-
+    private(set) lazy var didChangeCategoryPublisher = categoryInputView.$selectedCategoryType
     private(set) lazy var didChangeTitleTextPublisher = titleInputView.didChangeInputTextPublisher
     private(set) lazy var didChangeContentTextPublisher = contentInputView.didChangeInputTextPublisher
     private(set) lazy var didTapBarButtonPublisher = barButton.publisher(for: .touchUpInside)
@@ -25,16 +20,8 @@ final class MemoUpdateContentView: UIView {
     }
 
     private lazy var body = VStackView(spacing: 16) {
-        VStackView(spacing: 12, layoutMargins: .init(.horizontal, 16)) {
-            categoryView
-
-            VStackView(spacing: 4) {
-                categoryButton
-                categoryBorderView
-            }
-            .addConstraint {
-                $0.height.equalTo(40)
-            }
+        categoryInputView.configure {
+            $0.updateValue(modelObject)
         }
 
         titleInputView.configure {
@@ -58,10 +45,7 @@ final class MemoUpdateContentView: UIView {
         }
     }
 
-    private lazy var categoryView = createTitleView(.category)
-
-    private let categoryButton = MenuButton(type: .system)
-    private let categoryBorderView = BorderView()
+    private let categoryInputView = UpdateMenuInputView(.category)
     private let titleInputView = UpdateTextMultiInputView()
     private let contentInputView = UpdateTextMultiInputView()
 
@@ -76,7 +60,6 @@ final class MemoUpdateContentView: UIView {
 
         setupView()
         setupEvent()
-        setupMenu()
         setupBarButton()
     }
 
@@ -93,9 +76,7 @@ extension MemoUpdateContentView {
         if traitCollection.hasDifferentColorAppearance(comparedTo: previousTraitCollection) {
             super.traitCollectionDidChange(previousTraitCollection)
 
-            [barButton, categoryView].forEach {
-                $0.layer.borderColor = UIColor.primary.cgColor
-            }
+            barButton.layer.borderColor = UIColor.primary.cgColor
         }
     }
 }
@@ -112,114 +93,11 @@ extension MemoUpdateContentView {
 // MARK: - private methods
 
 private extension MemoUpdateContentView {
-    func createTitleView(_ type: MemoContentType) -> UIView {
-        let titleStackView = HStackView(spacing: 4) {
-            UIImageView()
-                .addConstraint {
-                    $0.size.equalTo(24)
-                }
-                .configure {
-                    $0.image = type.image
-                }
-
-            UILabel().configure {
-                $0.text = type.title
-                $0.textColor = .secondaryGray
-                $0.font = .boldSystemFont(ofSize: 16)
-            }
-
-            UIView()
-        }
-
-        return UIView()
-            .addSubview(titleStackView) {
-                $0.edges.equalToSuperview().inset(8)
-            }
-            .addConstraint {
-                $0.height.equalTo(40)
-            }
-            .apply(.inputView)
-    }
-
-    func setupCategory() {
-        var actions = [UIMenuElement]()
-
-        MemoInputCategoryType.allCases.forEach { categoryType in
-            actions.append(
-                UIAction(
-                    title: categoryType.title,
-                    image: categoryType.image,
-                    state: categoryType == selectedCategoryType ? .on : .off,
-                    handler: { [weak self] _ in
-                        self?.selectedCategoryType = categoryType
-                    }
-                )
-            )
-        }
-
-        categoryButton.configure {
-            var config = UIButton.Configuration.filled()
-            config.title = selectedCategoryType?.title
-            config.image = selectedCategoryType?.image?
-                .resized(size: .init(width: 24, height: 24))
-                .withRenderingMode(.alwaysOriginal)
-            config.baseForegroundColor = .primary
-            config.contentInsets = .init(top: 0, leading: 8, bottom: 0, trailing: 0)
-            config.imagePadding = 8
-            config.titleTextAttributesTransformer = .init { incoming in
-                var outgoing = incoming
-                outgoing.font = .boldSystemFont(ofSize: 16)
-                return outgoing
-            }
-            config.background.backgroundColor = .background
-            $0.configuration = config
-            $0.contentHorizontalAlignment = .leading
-            $0.showsMenuAsPrimaryAction = true
-            $0.menu = .init(
-                title: .empty,
-                options: .displayInline,
-                children: actions
-            )
-        }
-
-        categoryButton.$isShowMenu.sink { [weak self] in
-            self?.categoryBorderView.changeColor($0 ? .inputBorder : .primary)
-        }
-        .store(in: &cancellables)
-    }
-
     func setupEvent() {
         gesturePublisher().sink { [weak self] _ in
             self?.endEditing(true)
         }
         .store(in: &cancellables)
-    }
-
-    func setupMenu() {
-        guard let category = modelObject?.category else {
-            selectedCategoryType = .noSetting
-            return
-        }
-
-        switch category {
-        case .todo:
-            selectedCategoryType = .todo
-
-        case .technical:
-            selectedCategoryType = .technical
-
-        case .interview:
-            selectedCategoryType = .interview
-
-        case .event:
-            selectedCategoryType = .event
-
-        case .tax:
-            selectedCategoryType = .tax
-
-        case .other:
-            selectedCategoryType = .other
-        }
     }
 
     func setupBarButton() {
